rename + lint
