+++ internal/instance/store_test.go
